Handle invalid URIs in link token mapper
When we replace the tokens with links in the mapper, we do some checking
of the URL to sure it's an absolute GOV.UK URL. To do this, we pass the
string to `URI()` which will raise an exception if the URL is malformed.

There's not a whole lot we can do here as the URL that was originally
mapped came from some real GOV.UK content, so it's the source of truth
that's invalid.

Instead we can rescue the error and just return the URL as it was input
into the mapper. This will be presented to the user and we'll let the
browser handle it.

As an aside, the URL used in the spec here was the same format as the
one that originally triggered the error on production, with the angle
brackets being what triggered the exception. If we render that URL in
the browser it will actually strip the angle brackets and render a valid
link (in Chrome at least).