Preserve existing answer data when error is raised
If an unhandled error is raised when composing the answer, we currently
build an entire new answer object and return that from the composer.

This means we lose all existing data. For example if the error is raised
later on in the pipeline, we lose all the metrics and LLM responses from
earlier in the pipeline. This information can be useful when trying to
debug the reason the error was raised.

So now we use the existing answer if it exists, but overwrite the
relevant attributes so the user sees the canned response and we set the
right status.