Introduce new statuses for answers
Product are keen to remove the use of abort in our status collection. We
therefore have come up with a new status system (explained below). This
commit introduces the new status with TODO messages to remove the old
ones. The process we are taking to adding them is to deploy the code
with both old and new statuses in place and then backfill the old ones
to be the new ones prior to their removal.

---

Aggregate statuses

Modify our core group of aggregate statuses from: Success, Pending, Abort and Error

to: Answered, Pending, Unanswerable, Clarification, Guardrails, Banned and Error

Where

- Answered is a new name for Success to reflect that it is not the only type of success
- Unanswerable to cover a few cases - question routing that can’t be answered, no GOV.UK Content or LLM doesn’t answer
- Clarification to cover where question routing asks a user for information
- Guardrails for when any of the guardrails are failed (including forbidden terms)
- Banned for when a user is shadow banned

Full statuses

- success to be renamed to answered
- abort_question_routing to be changed to  unanswerable_question_routing or clarification (dependent on question routing label see below)
- abort_answer_guardrails to be changed to guardrails_answer
- abort_forbidden_terms to be changed to  guardrails_forbidden_terms
- abort_jailbreak_guardrails to be changed to guardrails_jailbreak
- abort_llm_cannot_answer to be changed to  unanswerable_llm_cannot_answer
- abort_no_govuk_content to be changed to unanswerable_no_govuk_content
- abort_question_routing_guardrails to be changed to guardrails_question_routing
- abort_question_routing_token_limit to be removed as something distinct and instead use unanswerable_question_routing or clarification - based on low likelihood of this occurring
- abort_user_shadow_banned to be changed to banned

no change to any error_* ones

Note: it is beneficial to our Metrics graphs to a single term as the prefix of statuses - please try maintain this convention