Remove abort_jailbreak_guardrails status
We want to remove the word "abort" from our enum values.

The enum value is already in place, we just need to update the places
where it's used.

We still need the `answer.status_abort_jailbreak_guardrails?` check in
the conditional as the old enum value will be around until we've
completely removed that enum.