Remove abort_question_routing_token_limit status
We've decided this status isn't really necessary as a) it's never come
up and b) we're not actually that bothered about distinguishing it from
an error.

But if the token limit is reached, we don't want to return the LLM
answer to the user as it'll be truncated. So we return a canned
response.