Update question routing statuses
We want to remove the word "abort" from our statuses.

Previously the question router would assign `abort_question_routing`
status to an answer if it wasn't classified as `genuine_rag`.

This commit changes it so that we assign either
`unanswerable_question_routing` or `clarification` as the status
depending on how the question was classified.

The actual value is stored against each classification in the config
file.