Stop automatically incrementing places after 22nd December
We now have a shut down date for GOV.UK Chat of the 23rd December with
signups being disabled in the morning. This stops adding new places for
both instant and delayed signups once the date is the 23rd, it has an
expectation that over the weekend of the 21st and 22nd we'll have built
up a backlog of places anyway as we won't have promotional materials
directing users to signup to GOV.UK Chat.