Remove access to the support page during downtime
We're closing the private beta soon. At which point we want to remove
access to the support page.

This updates the application layout to define the navigation links and
footer links and passes them into the header and footer components.

Rather than defining the nav links inside the header component.

One noteworthy thing here is that the db lookup to see if public access
is enabled is rescued. If the db is down and we don't resuce this then
we will serve a 500 error on all pages, even the static ones.