Skip check_chat_public_access for ErrorsController
This corrects a mistake that any error responses would get caught by the
check to see if public access is enabled and thus return the 409/503
response for any request to the app.

When chat is in a downtime state we expect only existing URLs to return
the downtime state and error pages to continue functioning as normal.