not_accepting_signups copy for pilot closing
These changes should not be merged until the 23rd December morning and
are in preparation for the closure of the GOV.UK Chat private beta.

This updates the copy to match the design provided in the Figma [1].
Both before and after designs in the figma had a back button, however we
didn't have one before. It seems of low consequence to add this in for
this design.

[1]: https://www.figma.com/design/OwyFINL9IsSRLlKjY0daoZ/GOV.UK-AI-team-2024%2F2025?node-id=52989-30219