Allow multiple basic auth credentials
This changes the approach to configuring basic auth so that we can use
env vars to create additional accepted credentials. This is something
that can't be achieved by the Rails http_basic_authenticate_with method
and thus we make use of lower level methods to achieve it. I
unfortunately had to copy and paste some of the Rails logic for checking
the password, including a cryptic requirement that the "and" condition
is not short-circuited.

I've set it up so that we can keep the existing BASIC_AUTH_USERNAME and
BASIC_AUTH_PASSWORD env vars, so we don't need to change any additional
configuration but can then add other ones (or replace those) with pairs
that have a distinguishing suffix. I expected that we'd only want a
handful of credentials max, otherwise this individual env var approach
would be a bit tedious to configure.

I've chosen to configure this with a runtime check on the environment
variables as it allows this to be tested easier.

The motivation for adding this was that we want to have 2 sets of basic
auth credentials in Heroku so that some can be shared with partners and
others with user research participants.