Replace answer feedback survey link
This removes a change made during the private beta to add a survey link
when a user submits feedback [1]. This doesn't apply an exact revert
though as I have not reinstated the JS feature that allowed a user to
hide feedback.

I haven't reinstated this feature as, given we turned it off to promote
a survey, it seems unnecessary.

[1]: https://github.com/alphagov/govuk-chat/pull/934