Remove app-landing-page-layout__flex-container class
This class appears to no longer be necessary as the page seems to render
the same with or without it. I believe it was previously necessary when
we had a quirksmode issue [1]

Doing something with this class has been on our to-do list for a while
as it uses a layout class outside of the layout (and thus is not
isolated) - deleting it is the simplest resolution to that problem.

[1]: https://github.com/alphagov/govuk-chat/pull/691