Set build secrets when building image
In order to build the Docker image when a new version is created, we
need to provide auth credentials when running `bundle install` as we'll
be pulling a private repo (govuk_chat_private).

The shared workflow `build-and-push-multiarch-image` in
govuk-infrastructure has been modified to allow this specific secret to
be passed in during the build stage, which in turn gets set when running
`docker build ...`.

The secret we're using here (`GOVUK_CI_GITHUB_API_TOKEN`) is a secret
available in the alphagov Github org and provides read access to all
repos in the org, so it'll be able to check out the private gem repo.

Using that secret, we concatenate it into a `username:password` string
which Bundler then uses to check out the private repo.