Set BUNDLER_GITHUB_TOKEN secret for Ruby workflows
These workflows have now been extended to allow this secret to be passed
in.

It will allow the workflows to install private gems during their Ruby
setup, which we'll need when we add the govuk_chat_private gem to this
repo.

For now the secret won't do anything, it's just laying the groundwork
for adding the private gem.