Set BUNDLE_GITHUB__COM env var in relevant workflows
We have a few steps in our workflows where we run the `ruby/setup-ruby`
action which installs Ruby and runs `bundle install`.

In order for Bundler to install gems that are hosted on private repos on
Github, it needs to be able to authenticate the request. It does this by
way of a `BUNDLE_GITHUB__COM` environment variable which is a
`username:password` string.

The password we're setting here is a secret available to all alphagov
repos which gives read access to all alphagov repos. So having this in
place will allow us to install the govuk_chat_private gem when it's
added to the Gemfile.