Update Dockerfile to handle private dependency
Now that we're using a gem that's hosted on a private Github repo
(govuk_chat_private), we need to modify the Dockerfile so it can
authenticate to Github when installing the dependencies.

The key change here is that we're mounting a secret value when running
`bundle install`. The secret is set in the CI workflow (see
.github/workflows/deploy.yml) where the Dockerfile is used to build the
image. Essentially that this does is set a `BUNDLE_GITHUB__COM`
environment variable for the `bundle install` step which allows Bundler
to pull down the private repo.

More information about secret mounting can be found in the docs[1].

Mounting secrets is a relatively new Docker Buildkit feature, so we need
to explicitly set the Dockefile syntax at the top of the file to version
1.

This commit also removes the `--platform=$TARGETPLATFORM` flag as this
is the default, so it's unnecessary to specify it.

The `--link` flag has been added to all calls to `COPY`. It uses hard
links instead of copying the file content into the image so it's faster
and reduces disk usage.

The `--frozen` flag is added to `bundle install` because without it the
build would fail as it tried to create a new `Gemfile.lock` file in a
read-only file system. It makes sense to use `--frozen` anyway as it ensures
that the Gemfile.lock file is not modified during the installation

[1] https://docs.docker.com/build/building/secrets/