Set Dependabot registry credentials
We've got an issue where Dependabot is failing because it can't pull
down the govuk_chat_private gem as it's behind a private repo.

Initially I thought ignoring the dependency would work (we can run
Dependabot ourselves on the govuk_chat_private repo), but it looks like
Dependabot will pull down all dependencies, then filter out the ones in
the ignore list[1].

So it seems like the way to fix this is to set some credentials so that
the gem can be pulled down.

This config is pulled from the Github docs[2]. I'm not sure if the
name needs to be "github-octocat" or if you can call it whatever you
want, but let's try what's in the docs first.

The secret referenced here is an org secret available to all of our
Github repos and gives read access to all of our repos, so it should
work.

[1] "Dependabot checks for all allowed dependencies and then filters out
any ignored dependencies or versions."
https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference#ignore--

[2] https://docs.github.com/en/enterprise-cloud@latest/code-security/dependabot/working-with-dependabot/configuring-access-to-private-registries-for-dependabot#git