Reference registry in Dependabot config
We have a custom registry defined in the Dependabot config, but
Dependabot throws an error if you don't actually reference it:

> The property '#/registries' includes the "github-octocat" registry which is not used in any of the configurations

Bundler uses multiple registries in this repo because we've got gems
coming from both RubyGems and Github now.

But from the docs[1], it specifies:

> Within the updates blocks, where you can use registries: "*" to tell Dependabot to use any or all of the registries you defined at the top level.

So hopefully by using the wildcard "*" it'll reference the custom
registry and install the govuk_chat_private gem from Github.

I've also changed the name of the registry from "github-octocat" to just
"github". It's still not clear whether this name really has any function
other than being a descriptor. "github-octocat" was used because that
was what was in the example in the docs[2]. But "github" makes much more
sense than the cryptically-named "github-octocat" so if we can use the
former, I think it makes more sense.

[1] https://docs.github.com/en/enterprise-cloud@latest/code-security/dependabot/working-with-dependabot/configuring-access-to-private-registries-for-dependabot#configuring-private-registries
[2] https://docs.github.com/en/enterprise-cloud@latest/code-security/dependabot/working-with-dependabot/configuring-access-to-private-registries-for-dependabot#git