Remove evaluation questions config file
We think it probably makes sense to remove this file from the codebase
ahead of open-sourcing. It doesn't contain any particularly sensitive
information but we'll err on the side of caution and get rid of it.

This file is used as input for generation the evaluation report via the
Rake task.

Instead, this commit changes the rake task so it accepts a parameter of
a path to an input file. The idea is that evaluators will supply their
own config file of questions.