Fix Dependabot config
The previous couple of commits were aimed at trying to tighten the
security of the Dependabot config by not allowing all dependencies to
allow insecure external code execution.

But it seems Dependabot only allows unique ecosystems in a file:

> Update configs must have a unique combination of 'package-ecosystem',
> 'directory', and 'target-branch'. Ecosystem 'bundler' has
> overlapping directories.

So we need to revert these changes to get Dependabot working again.