Set answer strategy via env var
Changes the approach to setting the answer strategy when a new
Question is created to now be configured via an env var. Currently
this defaults to `openai_structured_answer`, but allows for the
creation of an `ANSWER_STRATEGY` variable in `.env` which can be set
to `"claude_structured_answer"` in order to switch to using Claude
via Bedrock as an answer service.