Convert the OpenAIAnswer class to a more generic PipelineRunner
The `OpenAIAnswer` class has minimal code that is specific to OpenAI
so has been renamed to the more generic `PipelineRunner`. This class
will be used to execute the pipeline regardless of which answer
service is configured by the answer strategy.

Currently catches errors separately for both OpenAI and AWS, but
these may be abstracted in the future.