Add Claude structured answer composer
Creates a `StructuredAnswerComposer` class for Claude that makes
use of the AWS BedrockRuntime SDK to use Claude as an answer
service.

In order to force a JSON output from Claude it uses a tool call,
this is configured to always be used via the `tool_choice`
configuration option for the Bedrock request.