Export and configure AWS credentials
In order to start using Bedrock by making direct API calls to it, we
need to authenticate those requests with some credentials.

We can use dotenv here to load a specific file - in this case
`.env.aws.local` which will contain all the necessary env vars to allow
the Bedrock gem to authenticate.

There's a new script in the `bin` directory which uses gds-cli to
generate the credentials and write them to this file. A bit of filtering
on this output needs to be done though, as the CLI outputs a blob of
text like this:

```
unset AWS_CREDENTIALS_FILE;
unset AWS_DEFAULT_PROFILE;
unset AWS_PROFILE;
export AWS_REGION='eu-west-1';
...
```

Dotenv doesn't like the `unset` lines, and it includes the trailing
semicolon as part of the value in the `export` lines which makes the
credentials invalid.

So the `export-aws-credentials` script filters for just the lines
beginning with "export" and strips the trailing semicolon before writing
to the file.

The `bin/setup` script has been extended so it calls the above export
script. This just helps set everything up for developers using that
setup script. It's been modified to work with govuk-docker as well.