Extend LinkTokenMapper to handle anchor tags
We had an issue recently where the `LinkTokenMapper` returned an invalid
URL in an answer because the source URL was an anchor tag.

For context, the `LinkTokenMapper` will read some HTML and store the links
in a hash with a key of the URL and value of "link_x" where x is an
integer.

A method can then be called to pass in some markdown which contains
links in the form "link_x" and it rewrites those to the stored URLs.

However there's a bug when an anchor link is used. In this case, the
mapper will take a link like "#heading-1" and rewrite it to
"https://www.gov.uk/#heading-1".

This happens because we store the link's `href` in the mapping, then
when we come to replace the token with a link, it's seen as a relative
URL so we turn it into an absolute URL by prepending the website root to
it.

This works fine for relative URLs but not for anchor tags. In those
cases, we need to know the `exact_path` of the chunk we're using so we
can make it into a relative URL.

For example, if we have a chunk with the `exact_path` of
"/browse/business" and some HTML content of:

```
this is <a href="#anchor">an anchor</a>.
```

then when we come to replace the token with a link it needs to produce
the URL `https://www.gov.uk/browse/business#anchor`, rather than
`https://www.gov.uk/#anchor` which it does now.

To fix this we can pass the `exact_path` as a parameter in when mapping
the links to tokens. If the link is an anchor (i.e. starts with a "#")
then we store the exact path plus the anchor in the mapping.

It also refactors the functionality of the mapper a bit so that we
convert all URLs into absolute URLs before storing them in the map. This
makes things consistent and means we don't need to do any rewriting when
we pull the links back out.