Add question rephrasing to Claude pipeline
This adds a question rephrasing step to the Claude pipeline which
mirrors the existing OpenAI rephrasing.

Rather than just copying the existing implementation into the
`Claude` namespace and updating the API calls, instead we keep the
existing `Pipeline::QuestionRephraser` class and instead makes the
changes in that class to call provider-specific implementations instead.

The reason for this is that there's some logic in this step which is
applied before we actually make the API calls; namely that of:

* returning early if this is the first question in the conversation
* plucking out the last 5 relevant questions from the conversation to
  include in the prompt

So it makes sense to perform this logic in the
`Pipeline::QuestionRephraser` class before calling the provider APIs. To
that end, we now do the early return logic and find the relevant
`Question` records, then pass those records into provider-specific
classes.

Aside from that the implementation is relatively straightforward. We use
the same prompt and interpolate the same values into it.

In terms of the tests, I've added a new method: `bedrock_claude_text_response`. This
simply stubs a call to Bedrock with a specified text response. We didn't
have this previously as the structured answer call stubs the tool call
and tool call response, rather than a non-tool call text response.

Additionally, it optionally asserts on the "user" messages within the
`messages` array we pass to the Bedrock `converse` method. If the
`user_messages` argument is present (be it a string or a Regex), then an
element in the `messages` array must have a role of `"user"` and must
match the value in the argument, otherwise an error is raised.

This is similar to how the current OpenAI stubbing works, in that we
make assertions on the request body as well. It's trickier to make these
same assertions with the Bedrock library because we don't have the same
control over the request body.

With question rephrasing it's quite important that we make some
assertions on the request body (i.e. the messages) as some specs are
testing that the rephrased question is used in the prompt.