Check for signon access before chat offline
This modifies the ordering of these before actions so that if chat
requires signon for access then you will be authenticated for that
before the check for whether the application is offline or not.

The motivation for this is deintegrating chat from www.gov.uk and
switching back to requiring signon for chat access. As chat is currently
offline it surprised me that visiting
chat.integration.publishing.service.gov.uk showed the offline page
rather than signon. Thus I felt we should these round.