Add AWS env vars to Heroku manifest
Now that we support LLM calls to AWS bedrock this configures the
environment variables for Heroku review apps.

We could decide that we want to set up an env var here of
ANSWER_STRATEGY to force this to be Claude since most of our recent
changes will be Claude specific.