Prevent tests leaking prompts
We saw an issue recently where a spec failed on CI due to a prompt being
changed in govuk_chat_private and the RSpec output leaked the entire
content of the prompt.

We explicitly want to keep the prompts private, so this shouldn't
happen.

The main culprit is when we make an HTTP request which has been stubbed
out using Webmock, but the request made doesn't match the stubbed
request. Webmock then prints out both the stubbed body and the request
body. If the spec doesn't stub the prompt loading, then the prompt from
the private gem gets printed.

Really the only way to prevent this from happening is to never allow the
prompt config to be loaded in a spec, and to stub it out instead, so
that's what this commit does.

Mostly we just load a system prompt and a user prompt from the config
files, which are relatively straightforward to stub. Usually these can
be stubbed with a simple string, but sometimes some interpolation is
needed if the prompt contains interpolation tokens (e.g. "{input}").

This means it's a bit annoying if the prompt interpolation changes -
we'd need to remember to update the specs.

A lot of this will go away once we transition away from OpenAI, because
we don't use Webmock for stubbing with Bedrock.

Nothing sensitive is automatically displayed when testing against
Bedrock. Instead we have control over what we show in the test output,
and we can hide anything sensitive from being shown on CI, as is done in
this commit.

Initially I'd wanted to try to suppress the Webmock output globally if
in CI, but it doesn't appear to be possible.