Specify ordering of records in question rephraser spec
This spec is testing that we're pulling the last 5 most recent question
records and passing those to the provider rephraser class.

We're creating 6 records and asserting that the most 5 recent ones are
passed in, but that's not too obvious from the spec. If we create the
records with random timestamps, we can sort them and assert that we're
definitely pulling in the right ones.