Replace use of title component with heading
The title component has been removed from govuk_publishing_components [1]
and thus we need to replace it. The heading component with an xl
font-size is functionally equivalent.

I applied this by using Ashley's example [2], however I omitted setting
a margin_bottom of 8 as that matches the margin already added to
`govuk-heading-xl` and rendered identically.

In updating the tests I've switched from using a class based selector to
using element, h1. I've done this because it seems more semantic to
match on the element name since we only expect one h1 per response and
don't really mind what class it is in request tests.

[1]: https://github.com/alphagov/govuk_publishing_components/pull/4653
[2]: https://github.com/alphagov/collections/pull/3963/files