Use full path of .ruby-version file in Gemfile
Currently if you try to run `bundle exec ...` from anywhere other than
the root directory, you get an error:

```
There was an error parsing `Gemfile`: No such file or directory @ rb_sysopen - .ruby-version. Bundler cannot continue.
```

It's not a massive issue, but it's a nice little QOL improvement to
make to set the full path of the `.ruby-version` file.