Handle lists without links in step-by-step-nav
This relates to the bug fix in #119 where an issue with step-by-step-nav
parsing caused an unexpected system error. While that code fixes the
point in the application where there is a crash this fixes the creation
of links without a href.

This changes the handling of lists in a step-by-step-nav to not assume
that each item in a list is a link - as this is not a required attribute
in the schema [1].

As part of this change I've also added a space between link content and
context content since they are inline HTML elements.

[1]: https://github.com/alphagov/publishing-api/blob/main/content_schemas/formats/shared/definitions/_step_by_step_nav.jsonnet#L89-L109