Use template_mail when calling ApplicationMailer directly
A change in the `mail-notify` gem has caused this code to raise an error[1][2].

Using `ApplicationMailer.view_mail` will now try to find a view template for
the action:

```
 Failure/Error: super(template_id, options)

 ActionView::MissingTemplate:
   Missing template application_mailer/view_mail with "mailer".

   Searched in:
     * "application_mailer"
```

We don't really want to create a dummy view for this just to make the
tests pass and for our test notification Rake task. So instead we can
use a template mailer. From the docs:

> Template mailers only require the template ID from Notify and an to email
> address, all of the content for the email is managed in Notify

This has the downside for the test notification Rake task in that the
template specified in the env var needs to have a `subject` and `body`
placeholder in it, but given that this is a test notification job I
think that trade-off is OK to make.

Note: the `personalisation` param sent to the mailer is a hash of
placeholder names => values which will be interpolated in the template.

[1] https://github.com/dxw/mail-notify/commit/a18b5dc091b2b9207efd7cbe7d33ca95f532f8a5
[2] https://github.com/dxw/mail-notify/commit/462b7035086676c523acc844ae42da2240c479db