Rake task to generate a RAG answer
Data science are keen to have a means to evaluate how GOV.UK Chat
performs at the RAG level without the other components such as
guardrails and question routing.

To facilitate this I have written this Rake task which is expected to be
called from the Python evaluation framework [1] as a means to generate
the answers.

There was some key considerations in putting this together:

1. Where to put the code for generating the answer

I considered hardcoding the answer generation or
creating new answer strategies as means to implement this. I chose the
approach of hardcoding as I felt it offered a route to be less involved
in the AnswerComposition::Composer code and allowed skipping forbidden
words and other error handling.

Doing it this approach has probably increased the risk that this code
could be less maintained than the answer strategy since it is now in a
potentially surprisingly place.

2. Raising an error if an answer has an error status

For evaluation we ideally want all the questions to be answered. There
are however a couple of ways when they may not be answered - abort
statuses (such as unanswerable_no_govuk_content) and error statuses
(such as error_answer_service_error).

I decided to return data when it is an abort status as, following a chat
with Alessia, it looked like it'd be good to have that as data for a
report even though these wouldn't be explicitly evaluated.

Whereas for errors I thought it was best to raise here to make very
clear something had gone wrong. There's definitely room for improvement
with the error handling as just blowing up is a blunt instrument. I do
expect however that the majority of errors will be lack of an env var
and things that would indicate the whole process would fail so I think
erroring early is best.

3. Returning answer data via to_json

This decision is unlikely to be sufficient for long (we'll need extra
data like sources) but struck me as the simplest way to get data out
that we could use in evaluation quickly.

I expect this to be iterated.

[1]: https://github.com/alphagov/govuk-chat-evaluation-prototype/pull/12