Abort Claude answer generation when LLM cannot answer
This builds on the previous commit which extended the
'bedrock_claude_structured_answer_response' stub to allow for configuration
of the 'answered' attribute.

It updates the Claude structured_answer_composer to abort the pipeline when
the LLM has indicated that it cannot answer the question.

Much like the OpenAI structured_answer_composer, it sets the status of the
answer to 'unanswerable_llm_cannot_answer', sets the message to
'Answer::CannedResponses::LLM_CANNOT_ANSWER_MESSAGE' and assigns metrics.