Update AdminUser to SignonUser
We're going to have multiple types of users arriving from Signon.
Admin users who have full Signon accounts and API users who have an API
key use the new API we're building.

Due to this we need to rename the AdminUser model to SignonUser to reflect
the fact that it could be either type of user.