Add guardrail names to output guardrail evaluation task
The commit adds a list of guardrail names to the output of the
evaluation rake task for generating an output guardrail response. This
is in order to provide the exact number of guardrail definitions, as
well as their names to the evaluation tool which consumes this task (1).

Without providing these details they must be hardcoded in the evaluation
codebase, resulting in potential discrepancies if guardrail definitions
are updated.

1. https://github.com/alphagov/govuk-chat-evaluation-prototype/pull/24