Add blueprints required for the GET answer endpoint
This commit introduces the first blueprints we will use. The AnswerBlueprint
and GenericErrorBlueprint.

The AnswerBlueprint is responsible for serializing the answer object to
JSON before returning it to the client.

It only includes sources that have been used and uses the same logic that
chat does by using the Answer#group_used_answer_sources_by_base_path
method.

If answer feedback has been provided on the answer, it will be included
in response as laid out in the OpenAPI spec.

However, since this endpoint should only be used prior to feedback being given,
since it's a polling endpoint. I've still included it in the blueprint though
since it's present in the specification.

The GenericErrorBlueprint will be used when a non 422 exception is raised
to return a generic error message to the client.