Add GET answer endpoint
Since this is the first endpoint added i've done the normal:

- add route
- add controller
- add new endpoint to controller

This endpoint should respond in the following way:

- return an answer that adheres to the Answer object in the specification
if one is present with a 200
- return a 202 if no answer is present
- return a 404 if the conversation or question is not found

When we get a 404 i'd adhered to the same response we send in Chat i.e.

"Conversation not found" rather than using the error objects messsage.

I'll ensure the endpoint is behind auth and that the response adheres to
the OpenAPI specification in follow up commits.