Use Committee middleware to validate Chat API responses
This adds a new initializer for Committee that sets up the middleware
to validate the responses from the Chat API against the OpenAPI specification.

It is scoped to the Chat API V0 namespace and is configured to raise
errors if the response does not match the specification so we're alerted
via sentry rather than it failing silently.

I've set strict_reference_validation to true to ensure that our OpenAPI
schema is valid. This will be enforced in the next major release of
Committee.

You can read more about the configuration here:
https://github.com/interagent/committee?tab=readme-ov-file#committeemiddlewareresponsevalidation

I've also added a test to ensure that the middleware is being applied to
the endpoint.